gesture_control/requirements.txt
opencv-python
mediapipe
pyautogui
mss
pillow
PyQt6
